Extract the explain-pushdown rendering code
