+++ src/adapter/src/coord/sequencer/inner.rs
@@ -16,13 +16,17 @@ use std::time::{Duration, Instant};
 
 use anyhow::anyhow;
 use futures::future::BoxFuture;
+use futures::stream::FuturesOrdered;
 use futures::{future, FutureExt};
 use itertools::Itertools;
 use maplit::btreeset;
 use mz_adapter_types::compaction::CompactionWindow;
 use mz_cloud_resources::VpcEndpointConfig;
 use mz_controller_types::{ClusterId, ReplicaId};
-use mz_expr::{CollectionPlan, MirScalarExpr, OptimizedMirRelationExpr, RowSetFinishing};
+use mz_expr::{
+    CollectionPlan, MapFilterProject, MirScalarExpr, OptimizedMirRelationExpr, ResultSpec,
+    RowSetFinishing,
+};
 use mz_ore::collections::{CollectionExt, HashSet};
 use mz_ore::task::{self, spawn};
 use mz_ore::tracing::OpenTelemetryContext;
@@ -33,7 +37,7 @@ use mz_repr::adt::mz_acl_item::{MzAclItem, PrivilegeMap};
 use mz_repr::explain::json::json_string;
 use mz_repr::explain::{ExplainFormat, ExprHumanizer};
 use mz_repr::role_id::RoleId;
-use mz_repr::{Datum, Diff, GlobalId, Row, RowArena, RowIterator, Timestamp};
+use mz_repr::{Datum, Diff, GlobalId, IntoRowIterator, Row, RowArena, RowIterator, Timestamp};
 use mz_sql::ast::{
     CreateSubsourceStatement, Ident, MySqlConfigOptionName, UnresolvedItemName, Value,
 };
@@ -58,7 +62,9 @@ use mz_adapter_types::connection::ConnectionId;
 use mz_catalog::memory::objects::{
     CatalogItem, Cluster, Connection, DataSourceDesc, Secret, Sink, Source, Table, Type,
 };
+use mz_ore::cast::CastFrom;
 use mz_ore::instrument;
+use mz_persist_client::stats::SnapshotPartStats;
 use mz_sql::ast::AlterSourceAddSubsourceOption;
 use mz_sql::plan::{
     AlterConnectionAction, AlterConnectionPlan, CreateSourcePlanBundle, ExplainSinkSchemaPlan,
@@ -84,6 +90,7 @@ use mz_storage_client::controller::{
 };
 use mz_storage_types::connections::inline::IntoInlineConnection;
 use mz_storage_types::controller::StorageError;
+use mz_storage_types::stats::RelationPartStats;
 use mz_storage_types::AlterCompatible;
 use mz_transform::notice::{OptimizerNoticeApi, OptimizerNoticeKind, RawOptimizerNotice};
 use mz_transform::EmptyStatisticsOracle;
@@ -2201,6 +2208,99 @@ impl Coordinator {
                 )));
             }
         };
+    }
+
+    async fn render_explain_pushdown(
+        &mut self,
+        ctx: ExecuteContext,
+        as_of: Antichain<Timestamp>,
+        mz_now: ResultSpec<'static>,
+        imports: impl IntoIterator<Item = (GlobalId, MapFilterProject)>,
+    ) {
+        let explain_timeout = *ctx.session().vars().statement_timeout();
+
+        let mut futures = FuturesOrdered::new();
+        for (gid, mfp) in imports {
+            let catalog_entry = self.catalog.get_entry(&gid);
+            let full_name = self
+                .catalog
+                .for_session(&ctx.session)
+                .resolve_full_name(&catalog_entry.name);
+            let name = format!("{}", full_name);
+            let relation_desc = catalog_entry
+                .item
+                .desc_opt()
+                .expect("source should have a proper desc")
+                .into_owned();
+            let stats_future = self
+                .controller
+                .storage
+                .snapshot_parts_stats(gid, as_of.clone())
+                .await;
+
+            let mz_now = mz_now.clone();
+            // These futures may block if the source is not yet readable at the as-of;
+            // stash them in `futures` and only block on them in a separate task.
+            futures.push_back(async move {
+                let snapshot_stats = match stats_future.await {
+                    Ok(stats) => stats,
+                    Err(e) => return Err(e),
+                };
+                let mut total_bytes = 0;
+                let mut total_parts = 0;
+                let mut selected_bytes = 0;
+                let mut selected_parts = 0;
+                for SnapshotPartStats {
+                    encoded_size_bytes: bytes,
+                    stats,
+                } in &snapshot_stats.parts
+                {
+                    let bytes = u64::cast_from(*bytes);
+                    total_bytes += bytes;
+                    total_parts += 1u64;
+                    let selected = match stats {
+                        None => true,
+                        Some(stats) => {
+                            let stats = stats.decode();
+                            let stats = RelationPartStats::new(
+                                name.as_str(),
+                                &snapshot_stats.metrics.pushdown.part_stats,
+                                &relation_desc,
+                                &stats,
+                            );
+                            stats.may_match_mfp(mz_now.clone(), &mfp)
+                        }
+                    };
+
+                    if selected {
+                        selected_bytes += bytes;
+                        selected_parts += 1u64;
+                    }
+                }
+                Ok(Row::pack_slice(&[
+                    name.as_str().into(),
+                    total_bytes.into(),
+                    selected_bytes.into(),
+                    total_parts.into(),
+                    selected_parts.into(),
+                ]))
+            });
+        }
+
+        task::spawn(|| "explain filter pushdown", async move {
+            use futures::TryStreamExt;
+            let res = match tokio::time::timeout(explain_timeout, futures.try_collect::<Vec<_>>())
+                .await
+            {
+                Ok(Ok(rows)) => Ok(ExecuteResponse::SendingRowsImmediate {
+                    rows: Box::new(rows.into_row_iter()),
+                }),
+                Ok(Err(err)) => Err(err.into()),
+                Err(_) => Err(AdapterError::StatementTimeout),
+            };
+
+            ctx.retire(res);
+        });
     }
 
     #[instrument]
